OpenDSSC: fix incorrect truncations when using the `Abs` function; affects multiple components
The issue was probably introduced in the original translation from Pascal to C++.

First detected when investigating an issue with Capacitor when providing `R=0` and `XL!=0`, which should initialize R to `0.001*XL` but was resulting in a different value.

git-svn-id: https://svn.code.sf.net/p/electricdss/code@3938 d8739450-1e93-4ef4-a0af-c327d92816ff

+++ trunk/VersionC/PCElements/GICsource.cpp
@@ -494,7 +494,7 @@ void TGICSourceObj::GetVterminalForSource(int ActorID)
 		{
 			auto with0 = ActiveCircuit[ActorID]->Solution;
 			int stop = 0;
-			if(Abs((int) (with0->get_FFrequency() - SrcFrequency)) < EPSILON2)
+			if(Abs((with0->get_FFrequency() - SrcFrequency)) < EPSILON2)
 				Vmag = Volts;
 			else
 				Vmag = 0.0;
